create appointment

+++ src/main/java/com/pbl/backend/controller/UserController.java
@@ -24,13 +24,13 @@ public class UserController {
     }
 
     @GetMapping("/{userId}")
-    public ResponseEntity<UserDTO> getUserById(@PathVariable Integer userId){
+    public ResponseEntity<UserDTO> getUserById(@PathVariable Long userId){
         return ResponseEntity.ok(this.userService.getUserById(userId));
     }
 
     @PutMapping("/{userId}")
     public ResponseEntity<UserDTO> updateUser(
-            @PathVariable("userId") Integer id,
+            @PathVariable("userId") Long id,
             @Valid @RequestBody UserDTO userDTO) {
 
         userDTO.setUserId(id); // gán id từ path vào DTO
@@ -39,7 +39,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{userId}")
-    public ResponseEntity<?> deleteUser(@PathVariable Integer userId)
+    public ResponseEntity<?> deleteUser(@PathVariable Long userId)
     {
         this.userService.deleteUser(userId);
         return new ResponseEntity<ApiResponse>(new ApiResponse("User deleted Successfully", true), HttpStatus.OK);
