+++ src/main/java/com/pbl/backend/repository/DoctorRepository.java
@@ -4,5 +4,5 @@ import com.pbl.backend.model.Doctor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
-    Doctor findByUserId(Integer userId);
+    Doctor findByUserId(Long userId);
 }
