+++ src/main/java/com/pbl/backend/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    User findByUserId(Integer userId);
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByUserId(Long userId);
     User findByEmail(String email);
 }
